More work on project isolation
