+++ src/ConsoleApp/Program.cs
@@ -2,6 +2,11 @@
 {
     public static void Main(string[] arguments)
     {
+        Testing(1);
+    }
 
+    public void Testing()
+    {
+        
     }
 }
